MutatingOrNot

Documentation for MutatingOrNot.

MutatingOrNot.Allocators.dumb

MutatingOrNot.dryrun

MutatingOrNot.void

MutatingOrNot.Allocators.ArrayAllocator

MutatingOrNot.Allocators.SmartAllocator

MutatingOrNot.BasicDryRun

MutatingOrNot.BasicVoid

MutatingOrNot.DryRun

MutatingOrNot.Void

MutatingOrNot.Allocators.malloc

MutatingOrNot.Allocators.mfree

MutatingOrNot.Allocators.mfree

MutatingOrNot.has_dryrun

MutatingOrNot.similar

MutatingOrNot.similar!

MutatingOrNot.dryrun — Constant

dryrun is the only instance of the singleton type BasicDryRun. When passed as an output argument, it is meant to signal that one wants to allocate that output argument, but not to do actual work. See void and has_dryrun.

MutatingOrNot.void — Constant

void is the only instance of the singleton type BasicVoid <: Void. When passed as an output argument, it is meant to signal that this argument is not allocated, and needs to be. See Void, similar! and has_dryrun.

MutatingOrNot.BasicDryRun — Type

struct BasicDryRun <: DryRun end

See DryRun.

MutatingOrNot.BasicVoid — Type

struct BasicVoid <: Void end

See Void.

MutatingOrNot.DryRun — Type

abstract type DryRun<:Void end

Instances v of types subtyping DryDryn, especially dryrun, when passed as an output argument, are meant to signal that this argument needs to be allocated, but that no actual computation should take place. See has_dryrun.

MutatingOrNot.Void — Type

abstract type Void end

Instances v of types subtyping Void, especially void, when passed as an output argument, are meant to signal that this argument is not allocated, and needs to be. The aim is to implement both mutating and non-mutating styles in a single place, while facilitating the pre-allocation of output arguments before calling the mutating, non-allocating variant.

To facilitate this, the following behavior is implemented whenever v::Void

(; x, y, z) = v results in x==v etc.

x, y, z = v results in x==v etc.

@. v = expr returns @. expr

See also similar! and has_dryrun.

MutatingOrNot.has_dryrun — Method

has_dryrun(x) returns

true if x::DryRun,

any(has_dryrun, x) if x is a (named) tuple,

and false otherwise.

Use it to avoid computations when only allocations are desired. Example:

function f!(x, y, z)
    # allocations, if needed
    a = similar!(x.a, y)
    b = similar!(x.b, z)

    # early exit, if requested
    has_dryrun(x) && return (; a, b)

    # computations
    a = @. a = y*y
    b = @. b = exp(z)
    return (; a, b)
end

In the above example,

the special properties of x_::Void (see Void) are used

x = f!(void, y) is the non-mutating variant of f!

x = f!(dryrun, y) just allocates x, without performing actual work

x = f!(x, y) mutates the pre-allocated x (non-allocating)

MutatingOrNot.similar! — Method

xx = similar!(x, y...)

Convenience function that replaces more concisely:

if x::Void
    xx = similar(y...)
else
    xx = x
end

The goal is to allocate xx only when a pre-allocated x is not provided. similar(y...) defaults to Base.similar(y...) and similar(y) applies recursively to tuples and named tuples.

See also has_dryrun.

MutatingOrNot.similar — Method

x = similar(y...)

similar(y...) defaults to Base.similar(y...). Furthermore the single-argument similar(y) applies recursively to tuples and named tuples. Contrary to Base.similar, it is not possible to specify eltype or dims in this recursive variant.

MutatingOrNot.Allocators.dumb — Constant

The singleton dumb describes the simplest possible allocation strategy:

malloc(dumb, args...) == similar(args...)
mfree(dumb) = nothing

Especially, freeing memory is actually left to Julia's garbage collector.

MutatingOrNot.Allocators.ArrayAllocator — Type

Parent type for array allocators. An instance tmp::ArrayAllocator is to be used along the following pattern:

y = malloc(tmp, args...) # allocate temporary array y
... # do some computation with y
mfree(tmp, y) # `free` array y, in a sense depending on tmp

MutatingOrNot.Allocators.SmartAllocator — Method

smart = SmartAllocator()

Return a smart allocator smart, with the following behavior:

internally, smart maintains a store tracking previously allocated arrays, marked as either busy or free.

y = malloc(smart, args...) searches that store for an array with the appropriate eltype and shape. If this fails, it allocates one with similar(args...). Either way, y is marked as busy in the store.

mfree(smart, y) keeps y in the store and marks it as free, so that a later call to malloc can reuse it

any mfree must have a corresponding malloc

mfree(smart) marks all arrays in the store as free but does not actually free anything.

only empty!(smart) actually empties the store, allowing Julia's garbage collector to act.

Note of caution:

arrays from the store are reused only if they have the exact same eltype and shape as requested by malloc.

to avoid runaway memory usage, any malloc must have a corresponding mfree

this allocator is smart only if the eltype and shape requested via malloc belong to a small set of possibilities.

MutatingOrNot.Allocators.malloc — Function

y = malloc(tmp, args...)

Return array y, similarly to similar(args...). The allocator tmp may provide more or less sophisticated allocation strategies. See dumb and SmartAllocator.

MutatingOrNot.Allocators.mfree — Method

mfree(tmp)

Free allocator tmp. Whether anything is actually done depends on the allocator tmp. See dumb and SmartAllocator.

MutatingOrNot.Allocators.mfree — Method

mfree(tmp, y)

Free array y, which was previously allocated by malloc. Whether anything is actually done depends on the allocator tmp. See dumb and SmartAllocator.