MutatingOrNot

Documentation for MutatingOrNot.

MutatingOrNot.dryrun

MutatingOrNot.void

MutatingOrNot.BasicDryRun

MutatingOrNot.BasicVoid

MutatingOrNot.DryRun

MutatingOrNot.Void

MutatingOrNot.has_dryrun

MutatingOrNot.similar

MutatingOrNot.similar!

MutatingOrNot.dryrun — Constant

dryrun is the only instance of the singleton type BasicDryRun. When passed as an output argument, it is meant to signal that one wants to allocate that output argument, but not to do actual work. See void and has_dryrun.

MutatingOrNot.void — Constant

void is the only instance of the singleton type BasicVoid <: Void. When passed as an output argument, it is meant to signal that this argument is not allocated, and needs to be. See Void, similar! and has_dryrun.

MutatingOrNot.BasicDryRun — Type

struct BasicDryRun <: DryRun end

See DryRun.

MutatingOrNot.BasicVoid — Type

struct BasicVoid <: Void end

See Void.

MutatingOrNot.DryRun — Type

abstract type DryRun<:Void end

Instances v of types subtyping DryDryn, especially dryrun, when passed as an output argument, are meant to signal that this argument needs to be allocated, but that no actual computation should take place. See has_dryrun.

MutatingOrNot.Void — Type

abstract type Void end

Instances v of types subtyping Void, especially void, when passed as an output argument, are meant to signal that this argument is not allocated, and needs to be. The aim is to implement both mutating and non-mutating styles in a single place, while facilitating the pre-allocation of output arguments before calling the mutating, non-allocating variant.

To facilitate this, the following behavior is implemented whenever v::Void

(; x, y, z) = v results in x==v etc.

x, y, z = v results in x==v etc.

@. v = expr returns @. expr

See also similar! and has_dryrun.

MutatingOrNot.has_dryrun — Method

has_dryrun(x) returns

true if x::DryRun,

any(has_dryrun, x) if x is a (named) tuple,

and false otherwise.

Use it to avoid computations when only allocations are desired. Example:

function f!(x, y, z)
    # allocations, if needed
    a = similar!(x.a, y)
    b = similar!(x.b, z)

    # early exit, if requested
    has_dryrun(x) && return (; a, b)

    # computations
    a = @. a = y*y
    b = @. b = exp(z)
    return (; a, b)
end

In the above example,

the special properties of x_::Void (see Void) are used

x = f!(void, y) is the non-mutating variant of f!

x = f!(dryrun, y) just allocates x, without performing actual work

x = f!(x, y) mutates the pre-allocated x (non-allocating)

MutatingOrNot.similar! — Method

xx = similar!(x, y...)

Convenience function that replaces more concisely:

if x::Void
    xx = similar(y...)
else
    xx = x
end

The goal is to allocate xx only when a pre-allocated x is not provided. similar(y...) defaults to Base.similar(y...) and similar(y) applies recursively to tuples and named tuples.

See also has_dryrun.

MutatingOrNot.similar — Method

x = similar(y...)

similar(y...) defaults to Base.similar(y...). Furthermore the single-argument similar(y) applies recursively to tuples and named tuples. Contrary to Base.similar, it is not possible to specify eltype or dims in this recursive variant.